pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "harness-bench"
version = "0.1.0"
description = "Universal benchmarking framework for AI coding assistants"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Harness Bench Team" }
]
keywords = [
    "ai",
    "coding-assistant",
    "benchmark",
    "claude",
    "codex",
    "aider",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
aider = [
    "aider-chat>=0.50",
    "pexpect>=4.8",
]
openai = [
    "openai>=1.0",
]
anthropic = [
    "anthropic>=0.20",
]
all = [
    "harness-bench[dev,aider,openai,anthropic]",
]

[project.scripts]
harness-bench = "harness_bench.cli:main"

[project.urls]
Homepage = "https://github.com/harness-bench/harness-bench"
Documentation = "https://github.com/harness-bench/harness-bench#readme"
Repository = "https://github.com/harness-bench/harness-bench"

[tool.hatch.build.targets.wheel]
packages = ["src/harness_bench"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=harness_bench --cov-report=term-missing"
